requirements.txt
# StoryGenius AI小说创作平台依赖包
# Core dependencies
gradio>=4.0.0
requests>=2.28.0
python-dotenv>=1.0.0
tenacity>=8.2.0
Pillow>=9.0.0

# AI Provider APIs
openai>=1.0.0
anthropic>=0.25.0

# Book generation
EbookLib>=0.18

# Data handling
dataclasses-json>=0.5.0

# Logging and monitoring
# logging is built-in to Python, no need to install separately